fix(connlib): allow larger DNS responses (#5507)
Many name servers apply a limit as to how big a DNS response is allowed
to be to protect themselves against DoS attacks. Querying a domain with
large records can thus fail if all we have available is UDP. To mitigate
this, we configure every upstream / system DNS server to use UDP and TCP
and let hickory decide, when to use what.

In addition, we enable EDNS(0), an extension to the original DNS spec
that lifts several limits in terms of record sizes.