chore(connlib): implement some missing ICMP conversions (#5475)
So far, our packet translation only implemented the bare-minimum for
ICMP to work. There are a few things left that haven't been dealt with.
This PR adds additional conversions where it was easy.

There are still some left that require more elaborate mangling of the
packet, like updating pointer fields.