chore(clients): output human-formatted logs in addition to JSON (#5504)
Currently, the clients only send JSON formatted logs to the configured
log directory. These are very hard to read as a human because one has to
re-assemble the spans and fields that we use extensively in connlib's
logs.

With this patch, the logs are sent to two files: `.jsonl` as JSON
formatted and `.log` formatted in syslog format.