chore(connlib): only refresh DNS for connections that are in use (#5555)
With the current behavior after a connection stops being used it will
trigger a refresh DNS after every 30 seconds forever.

This can be bad for a gateway that could be handling more than thousands
of domain names.

This was prevented before by only setting `slated_for_refresh` when we
see the first packet, this was deprecated in favor of checking times in
the `handle_timeout`.

So the solution now is to check that the connection is being used
currently before triggering any DNS refresh.