fix(client/windows): fix upload speed by letting Wintun queue packets
Closes #5589. Refs #5571

Improves upload speeds on my Windows 11 VM from 2 Mbps to 10.5.

Sorted by decreasing speed:

- macOS host, Internet - 16 Mbps up
- Windows guest, Internet - 13 Mbps up
- macOS host, Firezone - 12 Mbps up
- Windows guest, Firezone, after this PR - 10.5 Mbps up
- Windows guest, Firezone, before this PR - 2 Mbps up

So it's not perfect but the worst part is fixed.

The slow upload speeds were probably a regression from #5571.
The MPSC channel only has a few spots in it, so if connlib doesn't pick up every
packet immediately (which would be impossible under load), we drop packets.
I measured 25% packet drops in an earlier commit.

I first tried increasing the channel size from 5 to 64, and that worked.

But this solution is simpler. I switch back to `blocking_send` so if connlib
isn't clearing the MPSC channel, Wintun will just queue up packets in its
internal ring buffers, and we aren't responsible for buffering.

Getting rid of `blocking_send` was a defense-in-depth thing to fix the
deadlock yesterday, but we still close the MPSC channel inside `Tun::drop`,
and I confirmed in a manual test that this will kick the worker thread out of
`blocking_send`, so the deadlock won't come back.